requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Pillow==10.0.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0 